pyproject.toml
[project]
name = "microbit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "microbit-stubs>=0.2.0",
    "uflash>=2.0.0",
]
